requirements.txt
asgiref==3.4.1
black==21.9b0
click==8.0.1
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-decouple==3.5
pytz==2021.3
regex==2021.9.30
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
